Modify structure of generated config file in UI
Instead of storing clinician data and division data separately, store data in "clinician" objects.
Each clinician object has an email property and a "divisions" property.
Each division property is a mapping between division name and min/max values for that division for that particular clinician.

This allows for easier viewing and managing of the data.

Also, add a warning when trying to create a new clinician with a name that is already used in data (case-sensitive).